requirements.txt
pandas
numpy
prophet
scikit-learn
tqdm
pyyaml
matplotlib
openpyxl
